Remove force touch preview
Apple deprecated the force touch capabilities and replace with UIContextMenuInteraction.
The issue with this new interface is that it does not have two states,
preview and commit that we were relying using force touch.
So instead I remove the force touch interaction all together and just
the long press action for all devices.