419 Allow camera with limited library permission
Previously, any attempt to add an image from the camera when you had only granted `PHAuthorizationStatusLimited` photo library permission resulted in an error message.

The example app handled that by directing users to go to settings and change their permission… but that’s not very obvious, after all, they’re not trying to add something from their library. Apple does not require library authorization for use of the camera.

This PR changes the `limited` permission to be handled identically to `authorized`, which removes the error message.