Guard against UInts too big for `addIndex` to handle
From Apple’s docs:

`index
Index to add. Must be in the range 0 .. NSNotFound - 1.`

This has caused crashes in the Woo app when we’ve had overflows from the `adjustedIndexForIndex` function’s arithmetic. Those overflows shouldn’t happen from here now, but this gives extra protection, and is closer to the previously existing check (that used != NSNotFound, but that was ineffective against a UInt as NSNotFound is integer max, not uint max.)